clearpath/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
ibm-watsonx-ai>=1.0.8
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
